chore(review): étape 6 — commentaires en anglais, tags de chantier et préfixes de logs supprimés
Passe purement textuelle (aucun changement de comportement) :
- Préfixes de phase supprimés des LOGS (~50 : É1-É10, B1-B9, C2-C9.6,
  D2a/b, A7…) — décision dev ; le log de transition combat gagne un
  préfixe sémantique « combat: » à la place.
- Sigles de chantier nus retirés des commentaires (C2/C3, B1-B6, R3-R7,
  H3/H8, 8.7c/e…) ; les ancres qualifiées par un doc restent
  (docs/RENDERING.md §5/§7, HORIZONTAL-PASS §H5…), comme [cpp-tuning]
  et U3-x/PG/D1-D7/V1-V4.
- Tout le français des commentaires traduit ; restent verbatim les
  identifiants de data ("Poney") et les titres de sections de docs
  français servant d'ancre (MEADOWS-PLAN §A, DUNGEON-GEN § « algorithme
  à tiroirs »).
- Fragments/faux corrigés : SceneEditor.hpp (ligne coupée),
  GridNavigator.hpp (c'est Z qui reste à 0, pas Y), LevelEditor.cpp
  (justification « rehash » fausse — drafts en uptr), InputActions
  (« F was free » historique).

+++ game/scenes/FxDirector.cpp
@@ -6,7 +6,7 @@
 #include "engine/core/Hash.hpp"
 #include "engine/render/FlyCamera.hpp"
 #include "engine/render/landscape/TerrainNoise.hpp" // MaterialWeights
-#include "game/SoundResolver.hpp" // C3: cue sounds
+#include "game/SoundResolver.hpp" // cue sounds
 
 namespace game {
 
@@ -44,8 +44,8 @@ void FxDirector::create(const data::FormDatabase& formsIn,
                 // Cosmetic seed: spot hash + a running counter (§8:
                 // presentation never touches the gameplay RNG).
                 const u32 seed =
-                    core::hashU32(static_cast<u32>(event.position.x * 73.0f) ^
-                                  (static_cast<u32>(event.position.z * 179.0f)
+                    core::hashU32(static_cast<u32>(static_cast<i32>(event.position.x * 73.0f)) ^
+                                  (static_cast<u32>(static_cast<i32>(event.position.z * 179.0f))
                                    << 8) ^
                                   ++spawnCounter);
                 sim->spawn(gameplay::toEmitterParams(*particles),
@@ -66,15 +66,15 @@ void FxDirector::create(const data::FormDatabase& formsIn,
                      cue->shakeDecay);
         }
         if (cue->sound.isValid() && sounds) {
-            // C3: the SoundForm resolver — weighted variant + jitter.
+            // The SoundForm resolver — weighted variant + jitter.
             // Its OWN always-incremented counter (review bug 7a: the
             // spawn counter only moved in the particle branch, freezing
             // sound-only cues on one variant) + the spot hash, mirroring
             // the particle seed; the emitter's scales soften sneaked
             // steps.
             const u32 seed =
-                core::hashU32(static_cast<u32>(event.position.x * 73.0f) ^
-                              (static_cast<u32>(event.position.z * 179.0f)
+                core::hashU32(static_cast<u32>(static_cast<i32>(event.position.x * 73.0f)) ^
+                              (static_cast<u32>(static_cast<i32>(event.position.z * 179.0f))
                                << 8) ^
                               ++soundCounter ^ 0xac00571cu);
             sounds->play(cue->sound, event.position, seed,
